Solution for LC21 Merge two sorted lists Added solution for https://leetcode.com/problems/merge-two-sorted-lists. Solution relies on iteration and a sentinel node to make coding easier.
Limitations
The sentinel node (i.e. creating a dummy head of the list that does not contain useful data but makes it easier to return the data we do care about) makes the code easier to write and follow at the cost of wasting some memory. Presumably, for typical usage this tradeoff is valid. However, in limited or performance-critical applications this may not be a feasible approach. In these cases, we might be better off writing more to reuse one of the lists (if modification of the argument lists is allowed) or writing code that only creates valid nodes.

Additionally, this solution assumes that the two argument lists are sorted.

Other approaches
This is written iteratively with a while loop. The same code within the while loop can be relatively easily be converted to a recursive call. The idea would be that for tempIterator.next the function would call itself again with the now modified lists as parameters. I chose not to go this approach because for arbitrarily large lists, this could produce a stack overflow with a large number of recursive calls.